% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_testing.R
\name{association_factor_testing}
\alias{association_factor_testing}
\title{ROMY factor association testing}
\usage{
association_factor_testing(
  Y,
  X,
  Z,
  num_factors = 3,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y = lm_learner,
  prediction_y = lm_predict,
  method = "double_cf",
  K = 5,
  split_ratio = c(0.3333, 0.3333, 1 - 0.3333 - 0.3333),
  parallel = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{Z}{A matrix containing the measurements for Z. No missing values allowed.}

\item{num_factors}{Number of factors to test, default is 3.}

\item{learner_x}{Prediction model function to learn prediction of X given Z. Defaul is linear regression.}

\item{prediction_x}{Prediction model to predict X given Z. Default is linear regression.}

\item{learner_y}{Prediction model function to learn prediction of Y given Z. Default is linear regression.}

\item{prediction_y}{Prediction model to predict Y given Z. Default is linear regression.}

\item{method}{Method for controlling cross-fitting. Options are 'double_cf', 'single_cf', or 'no_split'. Default is 'double_cf'.}

\item{K}{Value for K for the K fold cross fitting. Default is K=5.}

\item{split_ratio}{Ratio for splitting of training data for the two prediction tasks. Default is 0.5:0.5.}

\item{parallel}{Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation. Default is NULL.

return Dataframe containing the p-values for all tests.}
}
\description{
This function performs factor association testing between variables in Y and an element of X, adjusting for Z.
}
