% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_testing.R
\name{component_testing}
\alias{component_testing}
\title{ROMY component testing}
\usage{
component_testing(
  Y,
  X,
  Z,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y = lm_learner,
  prediction_y = lm_predict,
  K = 5,
  split_ratio = c(0.3333, 0.3333, 1 - 2 * 0.3333),
  num_factors = 3,
  parallel = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A list of matrices, containing the measurements for X as separate multi-dimensional components.}

\item{Z}{A matrix containing the measurements for Z. No missing values allowed.}

\item{learner_x}{Prediction model function to learn prediction of X given Z. Default is linear regression.}

\item{prediction_x}{Prediction model to predict X given Z. Default is linear regression.}

\item{learner_y}{Prediction model function to learn prediction of Y given Z. Default is linear regression.}

\item{prediction_y}{Prediction model to predict Y given Z. Default is linear regression.}

\item{K}{Value for K for the K fold cross fitting. Default is K=5.}

\item{split_ratio}{Ratio for splitting of training data for the two prediction tasks. Default is 0.5:0.5.}

\item{num_factors}{Number of factors to be extracted in Singular Value Decomposition. Default is 3.}

\item{parallel}{Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation. Default is NULL.

return Dataframe containing the p-values for all tests.}
}
\description{
This function performs component testing between Y and components of X, adjusting for Z.
}
