% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_testing.R
\name{covar_testing}
\alias{covar_testing}
\title{ROMY (co)variance testing}
\usage{
covar_testing(
  Y,
  X,
  Z,
  index_triples = NULL,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y_stage1 = lm_learner,
  prediction_y_stage1 = lm_predict,
  learner_y_stage2 = lm_learner,
  prediction_y_stage2 = lm_predict,
  method = "double_cf",
  K = 5,
  split_ratio = c(0.25, 0.25, 0.25, 0.25),
  parallel = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{Z}{A matrix containing the measurements for Z. No missing values allowed.}

\item{index_triples}{A dataframe with three columns containing the indices in X and Y to be tested, respectively. Default is NULL, leading to testing all possible triples.}

\item{learner_x}{Prediction model function to learn prediction of X given Z. Defaul is linear regression.}

\item{prediction_x}{Prediction model to predict X given Z. Defaul is linear regression.}

\item{learner_y_stage1}{Prediction model function to learn prediction of Y given X and Z in stage 1. Defaul is linear regression.}

\item{prediction_y_stage1}{Prediction model to predict Y given X and Z in stage 1. Defaul is linear regression.}

\item{learner_y_stage2}{Prediction model function to learn prediction of Y in stage 2. Defaul is linear regression.}

\item{prediction_y_stage2}{Prediction model to predict Y in stage 2. Defaul is linear regression.}

\item{method}{Method for controlling cross-fitting. Options are 'double_cf', 'single_cf, or 'no_split'. Default is 'double_cf'.}

\item{K}{Value for K for the K fold cross fitting. Default is K=5.}

\item{split_ratio}{Ratio for splitting of training data for the two prediction tasks. Default is 0.3333/0.3333/0.3333.}

\item{parallel}{Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation, default is NULL.

return Dataframe containing the p-values for all tests.}
}
\description{
This function tests for (co)variance effects between variables in Y and X, adjusting for Z.
}
