% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_testing.R
\name{.ace}
\alias{.ace}
\title{This functions performs the alternating condition expectation algorithm, the outcome is the interaction term.}
\usage{
.ace(interaction_term, X, Z, splits, training_part, learner, prediction)
}
\arguments{
\item{interaction_term}{Measurements for the interaction term.}

\item{X}{A matrix containing X.}

\item{Z}{A matrix containing the covariates Z.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{training_part}{The part of the training data that is used to train the prediction model. Default is 1.}

\item{learner}{Function pointer to prediction model (training). Default is 'lm_learner_simple'.}

\item{prediction}{Prediction model to predict Y given X and Z.}
}
\description{
This functions performs the alternating condition expectation algorithm, the outcome is the interaction term.
}
