% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{.cF_func}
\alias{.cF_func}
\title{F function as described in Appendix B of Chen 2022.}
\usage{
.cF_func(Y, X, inds_X, Gamma)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{inds_X}{List of indices for each column of Y that links covariates in X to the specific Y_m component.}

\item{Gamma}{Matrix Gamma}
}
\description{
F function as described in Appendix B of Chen 2022.
}
