% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.check_model}
\alias{.check_model}
\title{Prediction model checker}
\usage{
.check_model(learner, prediction)
}
\arguments{
\item{learner}{Function pointer for prediction model (training).}

\item{prediction}{Function pointer for prediction model (prediction).}
}
\description{
This function checks if the provided functions for training and prediction perform well on a simple task.
}
