% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{.covar_testing_sub_parallel}
\alias{.covar_testing_sub_parallel}
\title{This function computes the test statistic for (co)variance effects, using parallel computation}
\usage{
.covar_testing_sub_parallel(
  X,
  Xrs,
  Y,
  Z,
  col_X,
  col_Y,
  splits,
  BPPARAM,
  index_triples,
  learner_x,
  prediction_x,
  learner_y_stage1,
  learner_y_stage2,
  prediction_y_stage1,
  prediction_y_stage2,
  training1,
  training2,
  training3
)
}
\arguments{
\item{X}{A matrix containing the measurements for X.}

\item{Xrs}{Matrices containing the residuals for X.}

\item{Y}{A matrix containing the measurements for Y.}

\item{Z}{A matrix containing the measurements for Z.}

\item{col_X}{Column names for X.}

\item{col_Y}{Column names for Y.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation.}

\item{index_triples}{A dataframe with three columns containing the indices in X and Y to be tested, respectively.}

\item{learner_x}{Prediction model function to learn prediction of X given Z.}

\item{prediction_x}{Prediction model to predict X given Z.}

\item{learner_y_stage1}{Prediction model function to learn prediction of Y given X and Z in stage 1. D}

\item{learner_y_stage2}{Prediction model function to learn prediction of Y in stage 2.}

\item{prediction_y_stage1}{Prediction model to predict Y given X and Z in stage 1.}

\item{prediction_y_stage2}{Prediction model to predict Y in stage 2.}

\item{training1}{Training part for task 1.}

\item{training2}{Training part for task 2.}

\item{training3}{Training part for task 3.}
}
\description{
This function computes the test statistic for (co)variance effects, using parallel computation
}
