% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.create_splits}
\alias{.create_splits}
\title{K-fold}
\usage{
.create_splits(N, K, subs)
}
\arguments{
\item{N}{Overall sample size.}

\item{K}{Number of folds.}

\item{subs}{Number of sub samples.}
}
\description{
This function creates a partitioning of the samples for K-fold cross fitting, using subs sets for training.
}
