% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.create_splits_1sub}
\alias{.create_splits_1sub}
\title{K-fold with 1 training part}
\usage{
.create_splits_1sub(K, N)
}
\arguments{
\item{K}{Number of folds.}

\item{N}{Overall sample size.}
}
\description{
This function creates a partitioning of the samples for K-fold cross fitting, using 1 set for training.
}
