% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.create_splits_2subs}
\alias{.create_splits_2subs}
\title{K-fold with 2 training splits}
\usage{
.create_splits_2subs(K, N, split_ratio, single = FALSE)
}
\arguments{
\item{K}{Number of folds.}

\item{N}{Overall sample size.}

\item{split_ratio}{Fractions describing the allocation of training data for different training tasks.}

\item{single}{Boolean controlling single split or double splits. Default is FALSE.}
}
\description{
This function creates a partitioning of the samples for K-fold cross fitting, using 2 sets for training.
}
