% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.create_splits_3subs}
\alias{.create_splits_3subs}
\title{K-fold with 3 training splits}
\usage{
.create_splits_3subs(K, N, split_ratio)
}
\arguments{
\item{K}{Number of folds.}

\item{N}{Overall sample size.}

\item{split_ratio}{Fractions describing the allocation of training data for different training tasks.}
}
\description{
This function creates a partitioning of the samples for K-fold cross fitting, using 3 sets for training.
}
