% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_testing.R
\name{.dml_association_factor_testing}
\alias{.dml_association_factor_testing}
\title{This function computes the test statistics for double machine learning
based association testing, given the residual measurements and index pairs.
using parallel computations}
\usage{
.dml_association_factor_testing(
  Xrs,
  Y,
  Z,
  col_X,
  splits,
  factors,
  num_factors = 3
)
}
\arguments{
\item{Xrs}{Residual matrices for X.}

\item{Y}{Matrix of Y measurements.}

\item{Z}{Matrix of Z measurements.}

\item{col_X}{Column names for X.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{factors}{Factors estimated from data.}

\item{num_factors}{Number of factors to test, default is 3.}
}
\description{
This function computes the test statistics for double machine learning
based association testing, given the residual measurements and index pairs.
using parallel computations
}
