% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcit.R
\name{.dml_association_factor_testing_parallel}
\alias{.dml_association_factor_testing_parallel}
\title{This function computes the test statistics for double machine learning
based association testing, given the residual measurements and index pairs.
using parallel computations}
\usage{
.dml_association_factor_testing_parallel(
  Xrs,
  Y,
  Z,
  col_X,
  splits,
  factors,
  BPPARAM,
  training_y
)
}
\arguments{
\item{Xrs}{Residual matrices for X.}

\item{Y}{Matrix of Y measurements.}

\item{Z}{Matrix of Z measurements.}

\item{col_X}{Column names for X.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{factors}{Factors estimated from data.}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation.}

\item{training_y}{Training part for Y task.}
}
\description{
This function computes the test statistics for double machine learning
based association testing, given the residual measurements and index pairs.
using parallel computations
}
