% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit.R
\name{.dml_association_testing}
\alias{.dml_association_testing}
\title{This function computes the test statistics for double machine learning
based association testing, given the residual measurements and index pairs.}
\usage{
.dml_association_testing(Xrs, Yrs, col_X, col_Y, index_pairs)
}
\arguments{
\item{Xrs}{Residual matrices for X.}

\item{Yrs}{Residual matrices for Y.}

\item{col_X}{Column names for X.}

\item{col_Y}{Column names for Y.}

\item{index_pairs}{Pairs of X-Y to test.}
}
\description{
This function computes the test statistics for double machine learning
based association testing, given the residual measurements and index pairs.
}
