% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_testing.R
\name{.dml_association_testing_comp}
\alias{.dml_association_testing_comp}
\title{This function computes the test statistic for double machine learning based association testing, given the residual measurements.}
\usage{
.dml_association_testing_comp(Yrs, Xrs)
}
\arguments{
\item{Yrs}{Residual matrices for Y.}

\item{Xrs}{Residual matrices for X.}
}
\description{
This function computes the test statistic for double machine learning based association testing, given the residual measurements.
}
