% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{.estimate_K}
\alias{.estimate_K}
\title{Estimates K as described by Chen (2022).}
\usage{
.estimate_K(ev_values, c_delta, N, M)
}
\arguments{
\item{ev_values}{Eigenvalues from projected Gamma Matrix product}

\item{c_delta}{Constant for Delta Computation}

\item{N}{Sample Size}

\item{M}{Number of Phenotypes}
}
\description{
Estimates K as described by Chen (2022).
}
