% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar_testing.R
\name{.evaluate_models_covar}
\alias{.evaluate_models_covar}
\title{This function evaluates the trained prediction models for (co)variance testing.}
\usage{
.evaluate_models_covar(
  Y1,
  Y2,
  X,
  Z,
  model_data,
  prediction_y_stage1,
  prediction_y_stage2,
  prediction_x,
  variance
)
}
\arguments{
\item{Y1}{Measurements for Y1.}

\item{Y2}{Measurements for Y2.}

\item{X}{Measurements for X.}

\item{Z}{Measurements for Z.}

\item{model_data}{Fitted models.}

\item{prediction_y_stage1}{Prediction model to predict Y in stage 1.}

\item{prediction_y_stage2}{Prediction model to predict Y in stage 2.}

\item{prediction_x}{Prediction model to predict X.}

\item{variance}{Logical variable to indicate if Y1=Y2.}
}
\description{
This function evaluates the trained prediction models for (co)variance testing.
}
