% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_testing.R
\name{.factor_estimation}
\alias{.factor_estimation}
\title{This function performs the factor estimation}
\usage{
.factor_estimation(Y, X, Z, splits, training_part = 1, num_factors = 3)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{Z}{A matrix containing the measurements for Z. No missing values allowed.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.#' @param learner_x Prediction model function to learn prediction of X given Z. Defaul is linear regression.}

\item{training_part}{The part of the training data that is used to train the prediction model. Default is 1.}

\item{num_factors}{Number of factors to test, default is 3.}
}
\description{
This function performs the factor estimation
}
