% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_testing.R
\name{.get_individual_residuals}
\alias{.get_individual_residuals}
\title{Covariate adjustment and residual computation}
\usage{
.get_individual_residuals(
  Y,
  Z,
  ind_x,
  splits,
  factors,
  training_part = 3,
  learner = lm_learner_simple,
  prediction = lm_predict_simple,
  num_factors = 3
)
}
\arguments{
\item{Y}{A matrix containing the Y measurements.}

\item{Z}{A matrix containing the covariates Z.}

\item{ind_x}{Index of element of X to be tested.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{factors}{Factors estimated from data.}

\item{training_part}{The part of the training data that is used to train the prediction model. Default is 1.}

\item{learner}{Function pointer to prediction model (training). Default is 'lm_learner_simple'.}

\item{prediction}{Function pointer to prediction model (prediction). Default is 'lm_predict_simple'.}

\item{num_factors}{Number of factors to test, default is 3.}
}
\description{
This function performs the covariate adjustment for the factor association testing
}
