% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction_testing.R
\name{.get_linear_additive_residuals}
\alias{.get_linear_additive_residuals}
\title{This function computes residuals using a prediction model that does not incorporate interactions between X and Z.}
\usage{
.get_linear_additive_residuals(
  Outcome,
  X,
  Z,
  splits,
  training_part,
  learner,
  prediction
)
}
\arguments{
\item{Outcome}{A matrix containing the outcome measurements.}

\item{X}{A matrix containing X.}

\item{Z}{A matrix containing the covariates Z.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{training_part}{The part of the training data that is used to train the prediction model. Default is 1.}

\item{learner}{Function pointer to prediction model (training). Default is 'lm_learner_simple'.}

\item{prediction}{Prediction model to predict Y given X and Z.}
}
\description{
This function computes residuals using a prediction model that does not incorporate interactions between X and Z.
}
