% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_testing.R
\name{.get_quantile_res}
\alias{.get_quantile_res}
\title{Function to extract predicted residuals from quantile regression using quantreg R package}
\usage{
.get_quantile_res(fit, Y, Z, tau_value)
}
\arguments{
\item{fit}{Fitted object from quantile regression.}

\item{Y}{Outcome measurements.}

\item{Z}{Covariates.}

\item{tau_value}{Quantile value for quantile regression.}
}
\description{
Function to extract predicted residuals from quantile regression using quantreg R package
}
