% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_testing.R
\name{.get_quantile_residuals}
\alias{.get_quantile_residuals}
\title{Quantile function and residual computation}
\usage{
.get_quantile_residuals(Outcome, Z, splits, training_part = 2)
}
\arguments{
\item{Outcome}{A matrix containing the outcome measurements.}

\item{Z}{A matrix containing the covariates Z.}

\item{splits}{The splits of the samples determining the K-fold cross fitting.}

\item{training_part}{The part of the training data that is used to train the prediction model. Default is 2.}
}
\description{
Quantile function and residual computation
}
