% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.input_checks}
\alias{.input_checks}
\title{Input check}
\usage{
.input_checks(
  Y,
  X,
  Z,
  split_ratio,
  num_splits,
  nomissX,
  nomissY,
  K,
  parallel,
  BPPARAM
)
}
\arguments{
\item{Y}{Matrix Y.}

\item{X}{Matrix X.}

\item{Z}{Matrix Z.}

\item{split_ratio}{Split ratio for training.}

\item{num_splits}{Length of split ratio vector.}

\item{nomissX}{Logic value if missing values in X should be allowed.}

\item{nomissY}{Logic value if missing values in Y should be allowed.}

\item{K}{K value for cross-fitting.}

\item{parallel}{Logic value to indicate if parallel computing should be used.}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation.}
}
\description{
This function checks the input provided.
}
