% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{.is_categorical}
\alias{.is_categorical}
\title{helper function to determine if x has less or equal than 5 levels}
\usage{
.is_categorical(x)
}
\arguments{
\item{x}{Value vector.}
}
\description{
helper function to determine if x has less or equal than 5 levels
}
