% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.na_safe_matrix_product}
\alias{.na_safe_matrix_product}
\title{Matrix product with potential NA entries}
\usage{
.na_safe_matrix_product(A, B)
}
\arguments{
\item{A}{Matrix A.}

\item{B}{Matrix B.}
}
\description{
This function computes a matrix product that can deal with NA data.
}
