% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{.Q_func}
\alias{.Q_func}
\title{Q function as described in Appendix B of Chen 2022.}
\usage{
.Q_func(Y, X, inds_X, Gamma, Gamma_k, tau_k, lambda)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{inds_X}{List of indices for each column of Y that links covariates in X to the specific Y_m component.}

\item{Gamma}{Matrix Gamma}

\item{Gamma_k}{Matrix Gamma_k}

\item{tau_k}{Parameter tau_k}

\item{lambda}{Parameter lambda}
}
\description{
Q function as described in Appendix B of Chen 2022.
}
