% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{.romy_lf_algorithm}
\alias{.romy_lf_algorithm}
\title{Performs the proximal gradient descent algorithm as described by Chen 2022 (Appendix B).}
\usage{
.romy_lf_algorithm(
  Y,
  X,
  inds_X,
  epsilon,
  K_min,
  K_max,
  c_lambda,
  c_delta,
  max_iter = 20,
  eta = 0.5
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{inds_X}{List of indices for each column of Y that links covariates in X to the specific Y_m component.}

\item{epsilon}{Convergence threshold, default value is 0.01.}

\item{K_min}{Minimum number of latent factors, default value is 1.}

\item{K_max}{Maximum number of latent factors, default value is 5.}

\item{c_lambda}{factor for Lambda.}

\item{c_delta}{factor for Delta.}

\item{max_iter}{Maximum number of iterations, default value is 20.}

\item{eta}{Eta parameter, default is value is 0.5.}
}
\description{
Performs the proximal gradient descent algorithm as described by Chen 2022 (Appendix B).
}
