% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{.S_operator}
\alias{.S_operator}
\title{S operator as described in Appendix B of Chen 2022.}
\usage{
.S_operator(A, x)
}
\arguments{
\item{A}{Matrix A.}

\item{x}{Threshold x.}
}
\description{
S operator as described in Appendix B of Chen 2022.
}
