% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_testing.R
\name{.train_quantile_function}
\alias{.train_quantile_function}
\title{Function to train quantile regression using quantreg R package}
\usage{
.train_quantile_function(Y, Z, tau_value)
}
\arguments{
\item{Y}{Outcome measurements.}

\item{Z}{Covariates.}

\item{tau_value}{Quantile value for quantile regression.}
}
\description{
Function to train quantile regression using quantreg R package
}
