% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{lm_learner}
\alias{lm_learner}
\title{Training of linear regression model}
\usage{
lm_learner(Outcome, data, lambda = NULL)
}
\arguments{
\item{Outcome}{Outcome data.}

\item{data}{Covariate data for prediction.}

\item{lambda}{Value for lambda to ensure that Regression matrix is invertible.}
}
\value{
trained model.
}
\description{
This function performs training for a linear regression model, including interaction and squared terms.
}
