% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{lm_predict_simple}
\alias{lm_predict_simple}
\title{Prediction using simple linear regression model}
\usage{
lm_predict_simple(obj, data)
}
\arguments{
\item{obj}{Trained linear regression model.}

\item{data}{Covariate data for prediction.}
}
\value{
predicted Outcome values.
}
\description{
This function performs prediction using a linear regression model.
}
