% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{mlr3_gam_learner}
\alias{mlr3_gam_learner}
\title{Training of GAM model}
\usage{
mlr3_gam_learner(Outcome, data)
}
\arguments{
\item{Outcome}{Outcome data.}

\item{data}{Covariate data for prediction.}
}
\value{
trained model.
}
\description{
This function performs training for a GAM model.
}
