% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{mlr3_gam_predict}
\alias{mlr3_gam_predict}
\title{Prediction using GAM}
\usage{
mlr3_gam_predict(obj, data)
}
\arguments{
\item{obj}{Trained GAM model.}

\item{data}{Covariate data for prediction.}
}
\value{
predicted Outcome values.
}
\description{
This function performs prediction using a trained GAM model.
}
