% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{mlr3_lightgbm_predict}
\alias{mlr3_lightgbm_predict}
\title{Prediction using lightgbm}
\usage{
mlr3_lightgbm_predict(obj, data)
}
\arguments{
\item{obj}{Trained lightgbm model.}

\item{data}{Covariate data for prediction.}
}
\value{
predicted Outcome values.
}
\description{
This function performs prediction using a trained lightgbm model.
}
