% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit.R
\name{romy_cit}
\alias{romy_cit}
\title{ROMY-CIT}
\usage{
romy_cit(
  Y,
  X,
  Z,
  Z2 = NULL,
  index_pairs = NULL,
  transform_functions_x = NULL,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y = lm_learner,
  prediction_y = lm_predict,
  method = "no_split",
  K = 5,
  parallel = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{Z}{A matrix containing the measurements for Z, for adjusting Z. No missing values allowed.}

\item{Z2}{A matrix containing the measurements for Z2, for adjusting X. No missing values allowed, default is NULL. In this case, the computations use Z2=Z.}

\item{index_pairs}{A dataframe with two columns containing the pairs of indices in X and Y to be tested, respectively. Default is NULL, leading to testing all possible pairs.}

\item{transform_functions_x}{List of functions to transform the measurements for X. All transformations b(x) will be considered in testing. Default is NULL, implying b(x)=x.}

\item{learner_x}{Prediction model function to learn prediction of X given Z. Defaul is linear regression.}

\item{prediction_x}{Prediction model to predict X given Z. Default is linear regression.}

\item{learner_y}{Prediction model function to learn prediction of Y given Z. Default is linear regression.}

\item{prediction_y}{Prediction model to predict Y given Z. Default is linear regression.}

\item{method}{Method for controlling cross-fitting. Options are 'double_cf', 'single_cf', or 'no_split'. Default is 'no_split'.}

\item{K}{Value for K for the K fold cross fitting. Default is K=5.}

\item{parallel}{Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation. Default is NULL.

return Dataframe containing the p-values for all tests, including ACAT results (if more than 1 test per combination).}
}
\description{
This function performs association testing between variables in Y and X, adjusting for Z.
}
