% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov.R
\name{romy_cov}
\alias{romy_cov}
\title{ROMY-COV}
\usage{
romy_cov(
  Y,
  X,
  Z,
  Z2 = NULL,
  index_triples = NULL,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y_stage1 = lm_learner,
  prediction_y_stage1 = lm_predict,
  learner_y_stage2 = lm_learner,
  prediction_y_stage2 = lm_predict,
  K = 5,
  parallel = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{Z}{A matrix containing the measurements for Z, for adjusting Z. No missing values allowed.}

\item{Z2}{A matrix containing the measurements for Z2, for adjusting X. No missing values allowed, default is NULL. In this case, Z2=Z.}

\item{index_triples}{A dataframe with three columns containing the indices in X and Y to be tested, respectively. Default is NULL, leading to testing all possible triples.}

\item{learner_x}{Prediction model function to learn prediction of X given Z. Defaul is linear regression.}

\item{prediction_x}{Prediction model to predict X given Z. Defaul is linear regression.}

\item{learner_y_stage1}{Prediction model function to learn prediction of Y given X and Z in stage 1. Default is linear regression.}

\item{prediction_y_stage1}{Prediction model to predict Y given X and Z in stage 1. Default is linear regression.}

\item{learner_y_stage2}{Prediction model function to learn prediction of Y in stage 2. Default is linear regression.}

\item{prediction_y_stage2}{Prediction model to predict Y in stage 2. Default is linear regression.}

\item{K}{Value for K for the K fold cross fitting. Default is K=5.}

\item{parallel}{Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation, default is NULL.

return Dataframe containing the p-values for all tests.}
}
\description{
This function tests for (co)variance effects between variables in Y and X, adjusting for Z.
}
