% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcit.R
\name{romy_fcit}
\alias{romy_fcit}
\title{ROMY-FCIT}
\usage{
romy_fcit(
  Y,
  X,
  Z,
  Z2 = NULL,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y = lm_learner,
  prediction_y = lm_predict,
  K = 5,
  parallel = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{Z}{A matrix containing the measurements for Z, for adjusting Z. No missing values allowed.}

\item{Z2}{A matrix containing the measurements for Z2, for adjusting X. No missing values allowed, default is NULL. In this case, Z2=Z.}

\item{learner_x}{Prediction model function to learn prediction of X given Z. Defaul is linear regression.}

\item{prediction_x}{Prediction model to predict X given Z. Default is linear regression.}

\item{learner_y}{Prediction model function to learn prediction of Y given Z. Default is linear regression.}

\item{prediction_y}{Prediction model to predict Y given Z. Default is linear regression.}

\item{K}{Value for K for the K fold cross fitting. Default is K=5.}

\item{parallel}{Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized}

\item{BPPARAM}{BPPARAM object for BiocParallel parallel computation. Default is NULL.

return Dataframe containing the p-values for all tests.}
}
\description{
This function performs factor association testing between variables in Y and an element of X, adjusting for Z.
}
