% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf.R
\name{romy_lf}
\alias{romy_lf}
\title{ROMY-LF}
\usage{
romy_lf(Y, X, inds_X, epsilon = 0.01, K = 5, max_iter = 20, eta = 0.5)
}
\arguments{
\item{Y}{A matrix containing the measurements for Y.}

\item{X}{A matrix containing the measurements for X.}

\item{inds_X}{List of indices for each column of Y that links covariates in X to the specific Y_m component.}

\item{epsilon}{Convergence threshold, default value is 0.01.}

\item{K}{Number of latent factors, default value is 5.}

\item{max_iter}{Maximum number of iterations, default value is 20.}

\item{eta}{Eta parameter, default is value is 0.5.}
}
\description{
This function performs the conditional latent factor model computations
}
