% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{superlearner_binomial_learner}
\alias{superlearner_binomial_learner}
\title{Training of SuperLearner for binomial link functions}
\usage{
superlearner_binomial_learner(Outcome, data)
}
\arguments{
\item{Outcome}{Outcome data.}

\item{data}{Covariate data for prediction.}
}
\value{
trained model.
}
\description{
This function performs training for SuperLearner approach with binomial link function
}
