% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{superlearner_gaussian_predict}
\alias{superlearner_gaussian_predict}
\title{Prediction using SuperLearner (gaussian link)}
\usage{
superlearner_gaussian_predict(obj, data)
}
\arguments{
\item{obj}{Trained SuperLearner model.}

\item{data}{Covariate data for prediction.}
}
\value{
predicted Outcome values.
}
\description{
This function performs prediction using a trained SuperLearner model with gaussian link.
}
