% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{.create_no_split_data}
\alias{.create_no_split_data}
\title{no-cross-fitting-approach}
\usage{
.create_no_split_data(N)
}
\arguments{
\item{N}{Overall sample size.}
}
\description{
This function sets up the data such that evaluation and training data is the same.
}
